chore(test): split plg-onboarding test file for parallel throughput
The 8037-line plg-onboarding.test.js was the longest single file in
the parallel test suite by a wide margin. Under mocha --parallel each
worker processes a whole file, so the longest file sets the floor for
the entire suite. This was the next-biggest lever after the parallel
enablement in #2412.

Split into three files along natural describe boundaries:
- plg-onboarding-onboard.test.js (190 tests) — the 24 "onboard - *"
  describes
- plg-onboarding-status.test.js (39 tests) — getStatus, admin bypass,
  read-only admin denial, getAllOnboardings
- plg-onboarding-update.test.js (98 tests) — update, transitionStatus,
  admin onboarding management

Also adds a "Test File Size and Parallelism" guideline to CLAUDE.md so
future test authors split before a file becomes the floor again.

Test counts unchanged; no test logic changes. Local `npm test`
wall-clock drops meaningfully (absolute numbers vary by machine), and
the new floor is the largest split chunk rather than this one file.
